refactor: Remove dead code, add UML diagrams (class, use-case, sequence)
